+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Added
 - support RTL layout if locale.dir = rtf #4168
+- added estonian language
 - add more rtl configurations to languages that use it 
 
 ### Changed
+- update translations (15-08-2025)
 
 ### Fixed
 - share email contacts by email not by VCard #5364